Pulling ceph tests out into their own test suite
